+++ src/instrument/photon_noise.py
@@ -5,7 +5,7 @@ from domain.star import Star
 
 
 def apply_photon_noise_gauss_from_spectra2d(spectra_2d_exposure, channel: Channel, ctx: RunContext, star: Star):
-    distr = np.random.normal(loc=0.0, scale=1.0, size=spectra_2d_exposure.shape)
+    distr = np.random.normal(loc=0.0, scale=1.0, size=spectra_2d_exposure.shape).astype(np.float32)
     sigma = np.sqrt(np.clip(spectra_2d_exposure, 0, None))
     noise = distr * sigma
     return noise
